Add profile and tray managers; refactor UI components into separate modules
- Introduced ProfileManager for handling profile loading, saving, and management.
- Added TrayManager for managing system tray icon and menu interactions.
- Refactored UI components into dialogs.py and widgets.py for better organization and maintainability.
- Deprecated the old ui_components.py while maintaining backward compatibility through re-exports.
